+++ lib/src/index.ts
@@ -3,3 +3,4 @@ export * from "./scales";
 export * from "./series";
 export * from "./priceLine";
 export * from "./markers";
+export * from "./watermark";