+++ scripts/build-all.sh
@@ -24,10 +24,12 @@ cd "$REPO_ROOT/$EXAMPLES_PATH"
 # Set required env to build examples
 echo "Setting env variables..."
 
-export VITE_LIGHTWEIGHT_CHARTS_REACT_COMPONENTS_VERSION=$(jq -r .version $LIB_PACKAGE_JSON)
+export VITE_LIGHTWEIGHT_CHARTS_REACT_COMPONENTS_VERSION=$(jq -r .version $LIB_PACKAGE_JSON | sed 's/[^0-9.]//g')
 echo "lighweight-charts-components-version: $VITE_LIGHTWEIGHT_CHARTS_REACT_COMPONENTS_VERSION"
 
-export VITE_LIGHTWEIGHT_CHARTS_VERSION=$(jq -r '.dependencies["lightweight-charts"]'  $ROOT_PACKAGE_JSON)
+export VITE_LIGHTWEIGHT_CHARTS_VERSION=$(
+  jq -r '.dependencies["lightweight-charts"]' "$ROOT_PACKAGE_JSON" | sed 's/[^0-9.]//g'
+)
 echo "lightweight-charts-version: $VITE_LIGHTWEIGHT_CHARTS_VERSION"
 
 export VITE_GITHUB_URL=$(jq -r .repository $LIB_PACKAGE_JSON)