+++ lib/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2025-04-05
+### Feat
+- add WatermarkText and WatermarkImage components
+
 ## [0.1.1] - 2025-04-03
 ### Fix
 - fix empty markers on init issue